+++ colorchord2/hook.h
@@ -11,9 +11,9 @@ void HookKeyEvent( void (*KeyEvent)( void * v, int key, int down ), void * v );
 void UnhookKeyEvent( void (*KeyEvent)( void * v, int key, int down ), void * v );
 
 
-void SoundEventHappened( int samples, float * samps, int channel_ct, int is_out );
-void HookSoundInEvent( void (*SoundE)( void * v, int samples, float * samps, int channel_ct ), void * v, int is_out );
-void UnhookSoundInEvent( void (*SoundE)( void * v, int samples, float * samps, int channel_ct ), void * v, int is_out );
+void SoundEventHappened( int samples, short * samps, int channel_ct, int is_out );
+void HookSoundInEvent( void (*SoundE)( void * v, int samples, short * samps, int channel_ct ), void * v, int is_out );
+void UnhookSoundInEvent( void (*SoundE)( void * v, int samples, short * samps, int channel_ct ), void * v, int is_out );
 
 
 #endif
